Bananagrams

Description

A simple Bananagrams solver.

This module implements a heuristic for finding a valid Bananagrams layout given a list of characters and valid words. Note that for simplicity the heuristic only explores layouts where words in the same direction are not adjacent, and therefore might fail to find a solution even if one exists (this is however unlikely in realistic examples).

Documentation

data Hand

newHand :: [Text] -> Multiset Char -> Hand

Generates a new Hand for the given path and characters.