Bananagrams.Grid

Description

Bananagrams grid operations

Construction

data Grid s

A Bananagrams grid!

newGrid :: Int -> ST s (Grid s)

Generates a new grid of edge length 2*n+1, centered around 0.

Accessors

data Entry

A word entry inside a grid.

Constructors

Entry

Fields

entryText :: !Text

entryOrientation :: !Orientation

entryStart :: !YX

data Orientation

An entry's orientation.

Constructors

Horizontal

Vertical

orientationYX :: Orientation -> YX

gridEntries :: Grid s -> ST s [Entry]

Modification

Primitives

data Conflict

A conflict when adding an entry.

Constructors

Conflict

Fields

conflictYX :: !YX

conflictOldChar :: Char

conflictNewChar :: Char

setEntry :: Entry -> Grid s -> ST s (Maybe Conflict)

Attempts to add a new entry to the grid. Note that this method does *not* check that a word is valid. If the word is too long, this method will error.

unsetLastEntry :: Grid s -> ST s (Maybe Entry)

Removes the last added entry, or does nothing if the grid doesn't contain any entries.

Candidate locations

data Candidate

A candidate grid location.

Constructors

Candidate

Fields

candidateYX :: !YX

The candidate root coordinate, always already set in the grid.

candidateOrientation :: !Orientation

The orientation of the new word.

candidateChars :: !(Map Int Char)

All set characters in this candidate row or column, indexed from the root.

candidates :: Grid s -> ST s [Candidate]

Returns all candidate locations for adding new words to the grid. Note that word expansions are not considered valid candidates (e.g. T transforming CAT into CATS would not be returned).

Debugging

displayEntries :: [Entry] -> Either Conflict ByteString

Returns a human-readable representation of the entries.