Bananagrams

Description

A simple Bananagrams solver.

This module implements a heuristic for finding a valid Bananagrams layout given a list of characters and valid words. Note that for simplicity the heuristic only explores layouts where words in the same direction are not adjacent, and therefore might fail to find a solution even if one exists (this is however unlikely in realistic examples).

Documentation

solve :: Dictionary -> Multiset Char -> Maybe [Entry]

data Dictionary

newDictionary :: [Text] -> Dictionary

data Entry

A word entry inside a grid.

Constructors

Entry

Fields

entryText :: !Text

entryOrientation :: !Orientation

entryStart :: !YX

data Orientation

An entry's orientation.

Constructors

Horizontal

Vertical

displayEntries :: [Entry] -> Either Conflict ByteString

Returns a human-readable representation of the entries.