Bananagrams

Description

A simple Bananagrams solver.

This module implements a heuristic for finding a valid Bananagrams layout given a list of characters and valid words. Note that for simplicity the heuristic only explores layouts where words in the same direction are not adjacent, and therefore might fail to find a solution even if one exists (this is however unlikely in realistic examples).

Generate inputs

data Dictionary

A bag of allowed words.

newDictionary :: [Text] -> Dictionary

Generates a Dictionary from a list of words.

type Hand = Multiset Char

A convenience alias for the characters in hand (to be placed on the grid).

Find a solution

solve

Arguments

:: Severity
The minimum severity for which to log messages (to stderr).

-> Dictionary
The dictionary of allowed words.

-> Hand
The characters to assemble within the grid.

-> IO (Maybe [Entry])

Solves a game, assembling the characters into valid words as a grid.

data Severity

Constructors

Emergency

Alert

Critical

Error

Warning

Notice

Informational

Debug

Inspect the solution

data Entry

A word entry inside a grid.

Constructors

Entry

Fields

entryText :: !Text

entryOrientation :: !Orientation

entryStart :: !YX

data Orientation

An entry's orientation.

Constructors

Horizontal

Vertical

displayEntries :: [Entry] -> Either Conflict ByteString

Returns a human-readable representation of the entries.